Bonus US added

+++ sprint I/be_java_hisp_w15_g07/src/main/java/com/example/be_java_hisp_w15_g07/dto/request/NewPromoPostDTO.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class NewPromoPostDTO extends NewPostDTO {
 
     private Boolean has_promo;
-    private Float discount;
+    private Double discount;
 }
